fix(walker): use u64 for device and volume identifiers

+++ src/platform/mod.rs
@@ -15,7 +15,7 @@ use std::path::{Path, PathBuf};
 #[derive(Debug, Clone)]
 pub struct FileIdInfo {
     pub file_id: u64,
-    pub volume_serial: u32,
+    pub volume_serial: u64,
     pub number_of_links: u32,
 }
